+++ stringconv/number.go
@@ -1,4 +1,4 @@
-package strconv
+package stringconv
 
 import (
 	"fmt"
