+++ stringconv/number_test.go
@@ -1,4 +1,4 @@
-package strconv
+package stringconv
 
 import (
 	"reflect"
